add authRouter
