+++ api/api.js
@@ -3,13 +3,16 @@ import { Router } from "express";
 import messageRouter from "./messages/messageRouter.js";
 import bookingRouter from "./modules/bookings/bookingRouter.js";
 import deskRouter from "./modules/desks/deskRouter.js";
+import authRouter from "./modules/users/authRouter.js";
 import userRouter from "./modules/users/userRouter.js";
 
 const api = Router();
 
 api.use("/message", messageRouter);
 
-api.use("/auth/v1", userRouter);
+api.use("/v1/users", userRouter);
+
+api.use("/v1/auth", authRouter);
 
 api.use("/v1/desks", deskRouter);
